fix(kb): break the dox <-> dox-triage import cycle via a leaf staleness module
Tier A gate (biome lint .) flagged noImportCycles: dox-triage imports dox
(resolveRowPath) while dox imported dox-triage (readStaleness). Extract
AckRecord / STALENESS_VERSION / StalenessFile / readStaleness into the
dependency-free leaf packages/kb/src/staleness.ts; lint, triage, ack-on-edit
and query-time verdicts now share the one implementation through it. Barrel
re-exports unchanged for consumers (kb-extension, kb-plugin countStale).
