+++ packages/kb/src/cli.ts
@@ -6,7 +6,8 @@ import { existsSync, readFileSync } from "node:fs";
 import { join, resolve } from "node:path";
 import { frontmatterConfigHash, loadConfig, type ResolvedConfig, type ResolvedSource } from "./config.js";
 import { agentsChain, doxInit, doxLint } from "./dox.js";
-import { ackTargets, applyDecisions, buildWorkItems, readStaleness } from "./dox-triage.js";
+import { ackTargets, applyDecisions, buildWorkItems } from "./dox-triage.js";
+import { readStaleness } from "./staleness.js";
 import { evaluate, type GoldenItem } from "./eval.js";
 import { runIndexAtomic } from "./index-run.js";
 import { indexSource } from "./indexer.js";
