+++ packages/kb/src/index.ts
@@ -1,36 +1,40 @@
 // Public API barrel for @blackbelt-technology/pi-dashboard-kb (Phase 1 slice).
-
-export type { ChunkInput, ParseResult } from "./chunker.js";
-export { chunkMarkdown } from "./chunker.js";
-export type { FrontmatterConfig, GuardMode, KbConfig, RankingConfig, ReadDisciplineConfig, ResolvedConfig, SourceConfig } from "./config.js";
-export { DEFAULTS, frontmatterConfigHash, loadConfig, validateConfig } from "./config.js";
-export type { AgentsEntry, DoxInitPlan, DoxIssue, DoxLintResult } from "./dox.js";
-export { agentsChain, doxInit, doxLint, fallbackManifest, parseRowPaths, resolveRowPath } from "./dox.js";
-export type { AckRecord, StalenessFile } from "./dox-triage.js";
-export { ackTargets, applyDecisions, buildWorkItems, parseRows, readStaleness, replaceRowPurpose, STALENESS_VERSION } from "./dox-triage.js";
-export type { FacetKeyConfig, FmValue, ParsedFrontmatter, PropertyRow } from "./frontmatter.js";
-export { buildMeta, buildProperties, DEFAULT_FACET_KEYS, DEFAULT_SEARCHABLE_KEYS, parseFrontmatter, strictDate, strictNumber } from "./frontmatter.js";
-export type { IndexOptions, IndexSource, IndexStats } from "./indexer.js";
-export { indexSource } from "./indexer.js";
-export type { InitOptions, InitResult } from "./init.js";
-export { kbInit } from "./init.js";
-export type { RenderOpts } from "./render.js";
-export { renderHits } from "./render.js";
-export type { KbSourceKind, ResolveCtx, ResolvedSource, SourceResolver } from "./sources.js";
-export { classifyRef, resolveAll, resolverFor, sourceIdentity } from "./sources.js";
-export { SCHEMA_VERSION, SqliteFtsStore } from "./sqlite-store.js";
-export { canonicalSource, isTrusted, recordTrust, sourceHash } from "./trust.js";
 export type {
   Chunk,
   DocType,
   FileState,
   Filter,
   GraphEdge,
-  GraphNode,HitVerdict, 
+  GraphNode,
+  HitVerdict,
   KbHit,
   KbStore,
   SearchOpts,
-  StorePropertyRow,VerdictCounts, VerdictLabel 
+  StorePropertyRow,
+  VerdictCounts,
+  VerdictLabel,
 } from "./types.js";
+export { parseFrontmatter, buildMeta, buildProperties, strictNumber, strictDate, DEFAULT_SEARCHABLE_KEYS, DEFAULT_FACET_KEYS } from "./frontmatter.js";
+export type { FmValue, FacetKeyConfig, ParsedFrontmatter, PropertyRow } from "./frontmatter.js";
+export { renderHits } from "./render.js";
+export type { RenderOpts } from "./render.js";
+export { chunkMarkdown } from "./chunker.js";
+export type { ChunkInput, ParseResult } from "./chunker.js";
+export { SqliteFtsStore } from "./sqlite-store.js";
+export { indexSource } from "./indexer.js";
+export type { IndexSource, IndexOptions, IndexStats } from "./indexer.js";
+export { kbInit } from "./init.js";
+export type { InitOptions, InitResult } from "./init.js";
+export { loadConfig, validateConfig, DEFAULTS, frontmatterConfigHash } from "./config.js";
+export type { KbConfig, SourceConfig, RankingConfig, ResolvedConfig, FrontmatterConfig, GuardMode, ReadDisciplineConfig } from "./config.js";
+export { SCHEMA_VERSION } from "./sqlite-store.js";
+export { resolveAll, classifyRef, sourceIdentity, resolverFor } from "./sources.js";
+export type { KbSourceKind, ResolvedSource, ResolveCtx, SourceResolver } from "./sources.js";
+export { isTrusted, recordTrust, canonicalSource, sourceHash } from "./trust.js";
+export { agentsChain, doxInit, doxLint, fallbackManifest, parseRowPaths, resolveRowPath } from "./dox.js";
+export type { AgentsEntry, DoxInitPlan, DoxIssue, DoxLintResult } from "./dox.js";
+export { enrichHits, HASH_CAP_BYTES, COVERAGE_CAP_BYTES, SUBJECT_CAP } from "./verdict.js";
 export type { EnrichCtx, VerdictFs } from "./verdict.js";
-export { COVERAGE_CAP_BYTES, enrichHits, HASH_CAP_BYTES, SUBJECT_CAP } from "./verdict.js";
+export { ackTargets, applyDecisions, buildWorkItems, parseRows, replaceRowPurpose } from "./dox-triage.js";
+export { readStaleness, STALENESS_VERSION } from "./staleness.js";
+export type { AckRecord, StalenessFile } from "./staleness.js";
